Add files via upload
While I was an administrator in the Division of Science Management at Yeshiva University, I became the "go-to" person for any software or computer difficulties, due to my ability to quickly learn and teach others how to use technology. MaccaBuy, a customized version of the science procurement software Sciquest, was introduced by the University one month before I was hired.  When I arrived, I saw that my colleagues in the research laboratories were still having difficulty adjusting to the new purchasing software. Therefore, I came up with and created this quick guide to hand out.  My coworkers found it very helpful, and the number of requests for purchasing assistance from the management division greatly decreased.  I also used it to help each new hire to the science departments get acclimated more quickly.  Because of my guide, the new hires were able to catch on and use the software independently right away, which then made it easier for them to focus more on their research and less on